RAG/requirements.txt
python-dotenv
langchain
langchain_openai
pypdf
faiss-cpu